Impact definitions as a record (#298)
This is a huge one ... the aim is to get rid of the "dynamic"
`impacts.json` and have it in an elm file instead. This then doesn't
need all the loading and parsing, and should give us far less "passing
the impacts around", and less dealing with results and maybes.