Fix bogus github changelog decoding with ingredient editor. (#380)
The changelog page is broken because some of the github API response
author object fields are missing (ingredient editor commit data, cc
@ccomb). Note that it would be nice if the ingredient editor could have
an associated avatar image, but I'm not sure what would be required for
that (we don't want to create a github account for that). Gravatar
maybe? Anyway that's unrelated to this patch.


![image](https://github.com/MTES-MCT/ecobalyse/assets/41547/bf9ddb76-84d3-44bb-8aa3-83a02e64fb4f)