Data in backend (#587)
The staging branch already contains code to manage
[textile](https://github.com/MTES-MCT/ecobalyse/tree/staging/backend/textile)
data on the backend, but it is disabled.
This PR adds improvements to flush the work already done. It is still
disabled but reverting the last commit allows to enable it again. I'll
enable it in another branch and then work on something else.

- Store and modify textile json in the backend
- naming and fieldsets improvements on the forms
- allow to create an example by copying example json as is, from the
frontend to the backend