chore: store all ingredients in score_history (#645)
## :wrench: Problem

[User
Story](https://www.notion.so/Ajouter-l-ensemble-des-ingr-dients-alimentaires-score_history-06d2d2c87a7a42f38c3b815b5b8c6203?pvs=4)
Currently we only store the score of ~20 food examples products like
`Pizza royale`, `steak haché surgelé`, ...
These examples are very useful to check to understand the modification
in the method we do (for example, modifying the transport, the cooking
step,...)

But we are working on all ~250 ingredients : when removing azadirachtin
we are modifying the score of ~30 organic ingredients, but as these are
not in the examples. So we can't see the evolution in the score_history
dataviz app :(

## :cake: Solution

Store in score_history the score of all (~250) ingredients. But don't
add them to the list of food products examples as this would be
overwhelming for the end user.

## :rotating_light:  Points to watch / comments

- We need an invariant uuid for each food product stored. So I had to
create an uuid for each ingredient. To avoid storing this uuid in a file
somewhere, I recompute it each time using the invariant name (id) of the
food_product. Is that ok ?

## :desert_island: How to test

When `score_history` is launched, we should see the score of all
ingredients in the dataviz app.