fix: broken comparator charts on desynced cache data. (#706)
Sometimes when processes data are updated server side, the locally
cached client data are desynced, which triggers weird issues — most
notably in the comparator with missing chart segments, those
corresponding to missing data or linked objects.

This patch removes all use of the processes data cached in localStorage,
and loads detailed data on client app start when authenticated. That
way, we're always sure to start the client app with the latest available
data from the server.