feat(api,textile): make all country params optional. (#713)
It's been decided a while ago that we would handle the "default unknown
country" as a duplicate from one existing record in the country database
(India, as publicly documented in the methodology docs). While we could
challenge this decision, the requirements to provide countries for the
fabric, dyeing, ennoblement steps don't make sense anymore — because we
have a record for when the country is unknown.

This patch makes all the textile API country parameters optional and use
the "Unknown country" duplicate from the db as a fallback in all
computations.