chore: add PR template (#767)
## :wrench: Problem

PRs descriptions have different formats depending on the author of the
PR and don't always provide context and/or how to test.

## :cake: Solution

Setup a PR template so that every PR created starts with the same
structure.

See
https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository
for more information.

## :desert_island: How to test

We can't test it before merging it into `master`. Once merged into
`master`, the template should be the default content of a new PR
description.