fix: create object encrypted file for versions (#800)
## :wrench: Problem

We have an error in production when logged-in and going from old version
back to the `2.4.0` in the selector.


![image](https://github.com/user-attachments/assets/a63e8c10-f2a6-4905-a700-7cfcc630ade3)


## :cake: Solution

Add the encrypted file for objects when creating the version.


## :rotating_light:  Points to watch/comments

I’ve modified the already published version by adding the encrypted
`processes_impacts_object.json.enc` manually and redeployed the
production, so it should be working now. We don't need to land this
patch ASAP, it can wait for the next release.

## :desert_island: How to test

We will only be able to test it in the next release.