feat: Intégration Laine woolmark (#831)
## :wrench: Problem

We got new more robust LCI data for wool.

## :cake: Solution

Implement the new wool data: import into brightway, generate an export
(still using simapro data)

There is a new `external_db` option is the toplevel
`common.import_.import_simapro_csv` function to link the new activities
to an existing db. It's transmitted as is to the
`link_technosphere_by_activity_hash` strategy. This is because the new
wool data is created using EI 3.10 as a library in SimaPro, so we have
to import EI3.10 first then link to it.

## :desert_island: How to test

Get the new `wool.CSV.zip data`, import the wool using `make
import_ecoivent` then rebuild the image `make image` and do an export
with `make export_textile` and check there is no diff.
Check the new cool is one third below the previous one in terms of
impacts.

ecobalyse-data: laine_woolmark