chore(data): New ingredients (#814)
## :wrench: Problem

We're missing many ingredients. 

## :cake: Solution

Added ingredients and exported the impacts.

Along with the new ingredients, some other modifications that needed to
be done to get a more reliable result:
- The brightway search now makes sure there is only one result. This
forces to use more precise search terms to find a process. If it's
impossible to get only one results, setting the `search` field to the
exact `name` of the process forces to select this exact process.
- The full list of `activities.json` has also been adapted to this
behaviour

Since the ingredients editor is used when adding ingredients, it also
has been adapted and fixed:
- It has been adapted to this search behaviour, and the `local save`
button is disabled until the search term leads to a single result.
- It also now allows to set empty ecosystemic services, which was needed
while adding new ingredients
- Some select fields of the ecosystemic services have been translated
- A bug has been fixed regarding the new plural categories fields
(`process_category`)

## :desert_island: How to test

Check that the new ingredients on the review app are what is expected
(naming, impacts, identifier, etc.)


ecobalyse-private: ingredients

---------

Co-authored-by: Ingredient editor <ecobalyse@beta.gouv.fr>
Co-authored-by: Ingredient editor <paulboosz@gmail.com>