chore(data): Add irrigation to the Ecoinvent organic cotton (#832)
## :wrench: Problem

The organic cotton is a EI-modified one with only irrigation added. It's
impacts are hardcoded and not retrieved from either simapro or
brightway.

## :cake: Solution

Use the Ecoinvent organic cotton and add a strategy to add irrigation to
it during import.

## :rotating_light:  Points to watch/comments

This may only applies on brightway, as I'm unsure where to find the
modified organic cotton in SimaPro. It would be great to first merge
https://github.com/MTES-MCT/ecobalyse/pull/822

## :desert_island: How to test

Run the `make clean_data import_ecoinvent` then check in brightway / EI
3.9.1 and EI 3.10 that the organic cotton has an irrigation in the
technosphere.
Compare the new impact with the previous ones.

ecobalyse-data: organic_cotton