chore: readd score_history (#891)
## :wrench: Problem

With the deletion of the `data/` directory, the `score_history` was
removed too, making the CI build fail on `master`.

## :cake: Solution

Add the script back as it was in the `bin/` directory.

## :rotating_light:  Points to watch/comments

We should question the actual pertinence of this script and how we could
better maintain it, but it's not in the scope of this PR (the scope
being: make `master` green again).

## :desert_island: How to test

Go manually in Github actions
https://github.com/MTES-MCT/ecobalyse/actions, choose "Score History" on
the left and then run the workflow on this branch `fix/score-history`:
the workflow should be green.

---------

Co-authored-by: Nicolas Perriault <nicolas@perriault.net>