fix: add missing env and allow workflow dispatch for release creation (#892)
## :wrench: Problem

The current release workflow fails because it’s missing the transcrypt
key env variable. See
https://github.com/MTES-MCT/ecobalyse/actions/runs/12748629218

## :cake: Solution

Add the env variable and allow the possibility to run the workflow by
hand with a dispatch in case of failure.


## :desert_island: How to test

Hard to test as it is only testable when creating a release…

Once it has been approved, I’ll remove the already created tag and
restart the release process by hand using the workflow dispatch added
here.