feat: allow customize component element final mass (#959)
In the object simulator UI, add an number input for end users to update
final element weight in the material process unit (eg. `kg` for steel,
`m3` for wood, etc). Computation of the initially required material
amount should be transparently done and that result must be reflected in
the UI.