ci: use uv instead of pipenv for `score_history` (#1130)
## :wrench: Problem

The `score_history` script is [failing on
`master`](https://github.com/MTES-MCT/ecobalyse/actions/runs/15559004198/job/43806436083)
due to the move from `pip` to `uv`.

## :cake: Solution

Switch to `uv` and remove the need for the backend server in the CI as
we don’t use it in this script.

## :desert_island: How to test

Manually dispatch a new action `score_history` workflow on this branch,
it should be green like this one:
https://github.com/MTES-MCT/ecobalyse/actions/runs/15559884369/job/43809282316