chore: fix `scalingo` deploy (#1131)
## :wrench: Problem

Scalingo provides a deprecated Postgres DSN
`postgres:/…...?sslmode=prefer` should be
`postgresql+asyncpg:/……?ssl=prefer`.

## :cake: Solution

Replace the `DATABASE_URL` string in the start and create admins
scripts. Admin creation has been moved so that it is run on every deploy
with an upsert, even if it’s not a review app (useful for staging that
is not a review app).

## :desert_island: How to test

Deploy should be green.