fix: handle reusing outdated magic links. (#1141)
Users may get a disturbing "Unauthorized" error notification when
clicking a previously used magic link, this is because the backend API
responds with a generic 401 when a magic link token has already been
consumed.

In this specific situation and if the user is already signed in, don't
render any error message.