feat(bo): allow editing component scopes (#1118)
refs #1079 

### 🚨 Points to watch/comments

I've implemented the feature in the frontend without having the scopes
data attached to components in the backend. Right now when no `scopes`
attribute is attached to a component, it fallbacks to all scopes (food,
textile, veli, object), hence why the filtering feature in the admin ui
is not super useful until we have them.

---------

Co-authored-by: Vincent Jousse <vincent@jousse.org>