fix: 500 error on journal history (#1184)
## :wrench: Problem

When displaying the history of a component, the backend throws an error
500 when one of the users in the history is not the current user.
Moreover the format of the default 500 exception doesn’t follow the
problem details spec (`status_code` instead of `status`)
https://datatracker.ietf.org/doc/html/rfc9457

## :cake: Solution

Retrieve the related user when fetching journal entries and fix the
default exception format.


## :rotating_light:  Points to watch/comments

The two useless `.env` files of the `backend/` directory have been
removed too.

## :desert_island: How to test

Was fixed with TDD, so tests have been added to check that it works as
expected.

If you want to test by hand: create 2 different users, and modify the
same component with the two different users. You shouldn’t see any
error.