fix(backend): versions url rewriting (#1209)
## :wrench: Problem

- When browsing the `v6.0.0` version in the selector, a `/` is added to
the `clientUrl` causing http errors because of `//` in urls
- Versions under the `/versions/v6.0.0/` url for example are doing
HTTP calls to `/versions/v6.0.0/backend` instead of `/backend`

## :cake: Solution

- Use a regex to remove all trailing `/` in clientUrl
- Patch the v6.0.0
- Use a nginx rewrite rule for backend calls


## :rotating_light:  Points to watch/comments

I was unable to find how to manage the `.proxyrc.json` of Parcel to
handle URLs rewriting in a clean way, so I’ve hardcoded the version
number for now. Happy to receive some help here.

The best solution for me would be to use nginx locally like in
production to avoid those kind of problems and possible errors.

The `v6.0.0` release have been patched and deployed in production.

## :desert_island: How to test

In the review app, browse to the `v6.0.0` version in the selector and
try to login/register/see detailed impacts: it should work as expected.