refactor: let invalid cooking process be detected as a compilation error (#1289)
## :wrench: Problem

During migration to AGB 3.2 a test failure revealed a cooking process
hidden as a uuid in the code. That uuid needed to be changed but that
did not trigger a failure but a minor change in test results that could
have gone unseen and was not obvious.

## :cake: Solution

- Move that cooking process to the wellKnown processes so that it's a
needed process of the food sector.
- Remove the wildcard case (`_ ->`) so that a changed or missing cooking
triggers a compilation error.

## :desert_island: How to test

Just run the tests