feat(bo): add API to list processes (#1167)
## :wrench: Problem

fixes #1109 

We need a way to load the json processes file in the BO and an endpoint
to get the processes in order to display/use them in the frontend.

## :cake: Solution

Add a new table in the DB to store the processes.

## :rotating_light:  Points to watch/comments

I had to rewrite a custom auth middleware so that we can handle having a
different behavior for the same route, depending if the user is logged
in or not. It was needed to serve the detailed impacts or not.

At the same time I’ve implemented the ability to query the web API with
an ecobalyse token, it fixes #1173

## :desert_island: How to test

```bash
http http://localhost:1234/backend/api/processes 
```

Should list all the processes without the detailed impacts.

```bash
http http://localhost:1234/backend/api/processes  "Authorization: Bearer <json_web_token>"
```

and

```bash
http http://localhost:1234/backend/api/processes  "Authorization: Bearer <eco_api_token>"
```

Should list all the processes with the detailed impacts.

---------

Co-authored-by: Nicolas Perriault <nicolas@perriault.net>