fix: process tranforms loading
Using an inner join causes the transforms to not be loaded and a
MissingGreenlet error to append. It is because when using async IO lazy
loading does'nt work as expected and you need to be sure that your joins
are good otherwise a database operation is performed under the hood to
load the missing objects and it’s does’ne work in an async io context.