fix: remove pkill requirement (#1551)
## :wrench: Problem

When deploying the textile app on the new Scalingo 24 slim stack, we
have a deploy error because `pkill` is missing on the slim stack.

## :cake: Solution

Don’t use `pkill` and rely only on `kill` by remembering the PIDs.

## :desert_island: How to test

Deploy should be ok and we should have no errors in the scalingo logs.