chore: disable npm postinstall scripts (#1586)
## :wrench: Problem

As a security best practice, we want to avoid npm scripts automatically
running preinstall/postinstall scripts.

See
https://www.nodejs-security.com/blog/npm-ignore-scripts-best-practices-as-security-mitigation-for-malicious-packages
for background information.

## :cake: Solution

Make use of the `ignore-scripts` npm install option.
It would probably be enough to add it to the project’s .npmrc, but we
also explicitely add it to each npm install/ci command so that it’s
obvious everywhere that pre/postinstall scripts are suppressed

## :rotating_light:  Points to watch/comments

Transcrypt was configured in the package.json postinstall script; I’ve
added instructions to the README to configure it manually (it should
also be more reliable, as the postinstall was failing if the proper
envvar was not defined and exported beforehand)

## :desert_island: How to test

Stuff should keep working :sweat_smile: