feat: dump published components (#1572)
## :wrench: Problem

fixes #1390 

## :cake: Solution

Add a command to export published components, and add a workflow
dispatch on GH actions to test the automatic generation of the file.

## :rotating_light:  Points to watch/comments

Currently, the workflow is activated on every push here, I can/will
remove it before merging so that it’s only triggered by hand.

## :desert_island: How to test

Push an empty commit here, or try `uv run backend json dump-components`
locally with some components marked as `published` in the admin.